tests/wpt/web-platform-tests/css/css-fonts/parsing/font-synthesis-small-caps-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Module Level 4: parsing font-synthesis-small-caps with valid values</title>
<link rel="help" href="https://www.w3.org/TR/css-fonts-4/#font-synthesis-small-caps">
<meta name="assert" content="font-synthesis-small-caps supports the full grammar 'auto | none'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value('font-synthesis-small-caps', 'auto');
test_valid_value('font-synthesis-small-caps', 'none');
</script>
</body>
</html>
